Post index and show now visiable to guests

+++ app/views/posts/index.blade.php
@@ -4,7 +4,9 @@
 
 <h1>All Posts</h1>
 
-<p>{{ link_to_route('posts.create', 'Add new post') }}</p>
+@if (Auth::user())
+	<p>{{ link_to_route('posts.create', 'Add new post') }}</p>
+@endif
 
 @if ($posts->count())
 	<table class="table table-striped table-bordered table-hover">
@@ -25,13 +27,19 @@
 				<td>{{ $post->header }}</td>
 				<td>{{ $post->body }}</td>
 				<td>{{ $post->status }}</td>
-				<td>{{ link_to_route('posts.show', 'Show', array($post->id), array('class' => 'btn btn-primary')) }}</td>
-				<td>{{ link_to_route('posts.edit', 'Edit', array($post->id), array('class' => 'btn btn-info')) }}</td>
-				<td>
-					{{ Form::open(array('method' => 'DELETE', 'route' => array('posts.destroy', $post->id))) }}
-						{{ Form::submit('Delete', array('class'=> 'btn btn-danger')) }}
-					{{ Form::close() }}
-				</td>				
+
+				@if (Auth::user())
+					<td>{{ link_to_route('posts.show', 'Show', array($post->id), array('class' => 'btn btn-primary')) }}</td>
+					<td>{{ link_to_route('posts.edit', 'Edit', array($post->id), array('class' => 'btn btn-info')) }}</td>
+					<td>
+						{{ Form::open(array('method' => 'DELETE', 'route' => array('posts.destroy', $post->id))) }}
+							{{ Form::submit('Delete', array('class'=> 'btn btn-danger')) }}
+						{{ Form::close() }}
+					</td>
+				@else
+					<td>{{ link_to_route('posts.show', 'Show', array($post->id), array('class' => 'btn btn-primary')) }}</td>	
+				@endif	
+		
 			</tr>
 			@endforeach
